requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Natural Language Processing
nltk>=3.7
textblob>=0.17.1

# Social Media APIs
praw>=7.6.0  # Reddit API
tweepy>=4.12.0  # Twitter API

# Web scraping (backup)
requests>=2.28.0
beautifulsoup4>=4.11.0

# Data processing
python-dateutil>=2.8.0
pytz>=2022.1

# Configuration
python-dotenv>=0.19.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.15.0

# Statistical analysis
statsmodels>=0.13.0
scikit-learn>=1.1.0

# Progress bars
tqdm>=4.64.0

# Async operations (for Twitter API)
aiohttp>=3.8.0

# Data validation
cerberus>=1.3.4
